#567: refactor to work with independent components

+++ mdl/src/main/herd/scripts/installDBForHerd.sh
@@ -262,22 +262,22 @@ if [ "${herdRollingDeployment}" = "false" ] ; then
         execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"INSERT INTO dmrowner.user_tbl VALUES ('${admin_user_url}', 'USER', 'ADMIN', current_timestamp, current_timestamp, 'SYSTEM', 'SYSTEM', '${PGUSER}', 'Y', 'Y');\""
 
 
-        # Add Herd namespace-auth SNS topic configuration
-        #todo only execute when bdsql auth is deployed
+        # Add Herd namespace-auth SNS topic configuration(only execute when bdsql auth is deployed)
         sns_arn=$(aws ssm get-parameter --name /app/MDL/${mdlInstanceName}/${environment}/Resources/SNS/UserNamespaceChgTopicArn --region ${region} --output text --query Parameter.Value)
-        echo "Inserting SNS config for user-namespace authorization status changes. SNS Arn: ${sns_arn}"
-        execute_cmd "sed -i \"s/{{SNS_TOPIC_ARN}}/${sns_arn}/g\" ${deployLocation}/sql/nsAuthSnsConfig.sql"
-        execute_cmd "sed -i \"s/{{ENVIRONMENT}}/${environment}/g\" ${deployLocation}/sql/nsAuthSnsConfig.sql"
-        execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -f ${deployLocation}/sql/nsAuthSnsConfig.sql"
+        if [ -n "$sns_arn" ]; then
+            echo "Inserting SNS config for user-namespace authorization status changes. SNS Arn: ${sns_arn}"
+            execute_cmd "sed -i \"s/{{SNS_TOPIC_ARN}}/${sns_arn}/g\" ${deployLocation}/sql/nsAuthSnsConfig.sql"
+            execute_cmd "sed -i \"s/{{ENVIRONMENT}}/${environment}/g\" ${deployLocation}/sql/nsAuthSnsConfig.sql"
+            execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -f ${deployLocation}/sql/nsAuthSnsConfig.sql"
 
-        # ensure that jms publishing is enabled
-        execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"DELETE FROM cnfgn WHERE cnfgn_key_nm = 'jms.listener.enabled';\""
-        execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"INSERT INTO cnfgn VALUES ('jms.listener.enabled','true', NULL);\""
+            # ensure that jms publishing is enabled
+            execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"DELETE FROM cnfgn WHERE cnfgn_key_nm = 'jms.listener.enabled';\""
+            execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"INSERT INTO cnfgn VALUES ('jms.listener.enabled','true', NULL);\""
 
-        # ensure sqs/sns publishing is enabled
-        execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"DELETE FROM cnfgn WHERE cnfgn_key_nm = 'herd.notification.sqs.enabled';\""
-        execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"INSERT INTO cnfgn VALUES ('herd.notification.sqs.enabled','true', NULL);\""
-
+            # ensure sqs/sns publishing is enabled
+            execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"DELETE FROM cnfgn WHERE cnfgn_key_nm = 'herd.notification.sqs.enabled';\""
+            execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"INSERT INTO cnfgn VALUES ('herd.notification.sqs.enabled','true', NULL);\""
+        fi
     fi
 
     # Configuration for Elasticsearch
@@ -303,12 +303,14 @@ if [ "${herdRollingDeployment}" = "false" ] ; then
     execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"INSERT INTO dmrowner.ec2_od_prcng_lk VALUES (5895, '${region}', 'm4.2xlarge', 0.40000, now(), 'SYSTEM', now(), 'SYSTEM');\""
 
     # Add Herd business object data change SNS topic configuration
-    sns_arn=$(aws ssm get-parameter --name /app/MDL/${mdlInstanceName}/${environment}/Resources/SNS/BusinessObjectDataChgTopicArn --region ${region} --output text --query Parameter.Value)
-    echo "Inserting SNS config for business object data status changes. SNS Arn: ${sns_arn}"
-    execute_cmd "sed -i \"s/{{SNS_TOPIC_ARN}}/${sns_arn}/g\" ${deployLocation}/sql/bzObjChgSnsConfig.sql"
-    execute_cmd "sed -i \"s/{{ENVIRONMENT}}/${environment}/g\" ${deployLocation}/sql/bzObjChgSnsConfig.sql"
-    execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -f ${deployLocation}/sql/bzObjChgSnsConfig.sql"
-    execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"UPDATE cnfgn set cnfgn_value_ds = 'true' where cnfgn_key_nm='herd.notification.sqs.enabled';\""
+    glue_sns_arn=$(aws ssm get-parameter --name /app/MDL/${mdlInstanceName}/${environment}/Resources/SNS/BusinessObjectDataChgTopicArn --region ${region} --output text --query Parameter.Value)
+    if [ -n "$glue_sns_arn" ]; then
+        echo "Inserting SNS config for business object data status changes. SNS Arn: ${glue_sns_arn}"
+        execute_cmd "sed -i \"s/{{SNS_TOPIC_ARN}}/${sns_arn}/g\" ${deployLocation}/sql/bzObjChgSnsConfig.sql"
+        execute_cmd "sed -i \"s/{{ENVIRONMENT}}/${environment}/g\" ${deployLocation}/sql/bzObjChgSnsConfig.sql"
+        execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -f ${deployLocation}/sql/bzObjChgSnsConfig.sql"
+        execute_cmd "psql --set ON_ERROR_STOP=on --host ${herdDatabaseHost} --port 5432 -c \"UPDATE cnfgn set cnfgn_value_ds = 'true' where cnfgn_key_nm='herd.notification.sqs.enabled';\""
+    fi
 fi
